feat: updated the version of node and react and used trustvc renderer
