+++ package.json
@@ -5,18 +5,19 @@
   "main": "build/index.js",
   "types": "build/index.d.ts",
   "dependencies": {
-    "@emotion/core": "10.0.27",
-    "@emotion/styled": "10.0.27",
-    "@govtechsg/decentralized-renderer-react-components": "2.0.3",
-    "@hot-loader/react-dom": "16.10.2",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@fortawesome/fontawesome-svg-core": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@trustvc/decentralized-renderer-react-components": "file:../decentralized-renderer-react-components",
     "bootstrap": "^4.3.1",
     "date-fns": "^1.30.1",
     "debug": "4.1.1",
     "lodash": "^4.17.15",
     "moment-timezone": "^0.5.26",
-    "react": "16.10.2",
-    "react-dom": "16.10.2",
-    "react-hot-loader": "4.12.19"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "scripts": {
     "build": "NODE_ENV=production webpack --progress --mode production",
@@ -31,48 +32,58 @@
     "semantic-release": "semantic-release"
   },
   "devDependencies": {
-    "@babel/core": "7.8.4",
-    "@babel/preset-env": "7.8.4",
-    "@babel/preset-react": "7.8.3",
-    "@babel/preset-typescript": "7.8.3",
-    "@emotion/babel-preset-css-prop": "10.0.27",
+    "@babel/core": "^7.27.0",
+    "@babel/preset-env": "^7.27.0",
+    "@babel/preset-react": "^7.26.3",
+    "@babel/preset-typescript": "^7.26.0",
+    "@emotion/babel-preset-css-prop": "^11.12.0",
     "@storybook/addon-actions": "5.3.9",
     "@storybook/addon-docs": "5.3.9",
     "@storybook/addons": "5.3.9",
     "@storybook/react": "5.3.9",
-    "@testing-library/react": "9.4.0",
-    "@types/debug": "4.1.5",
-    "@types/jest": "25.1.1",
+    "@testing-library/react": "^16.1.0",
+    "@types/debug": "4.1.12",
+    "@types/jest": "^29.5.14",
     "@types/lodash": "^4.14.149",
     "@types/moment-timezone": "^0.5.12",
-    "@types/react": "16.9.19",
-    "@typescript-eslint/eslint-plugin": "2.18.0",
-    "@typescript-eslint/parser": "2.18.0",
-    "babel-jest": "25.1.0",
-    "babel-loader": "8.0.6",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
+    "@typescript-eslint/eslint-plugin": "^7.5.0",
+    "@typescript-eslint/parser": "^7.5.0",
+    "assert": "^2.1.0",
+    "babel-jest": "^29.7.0",
+    "babel-loader": "^9.2.1",
     "brotli-webpack-plugin": "1.1.0",
-    "compression-webpack-plugin": "3.1.0",
-    "copy-webpack-plugin": "^5.1.1",
-    "css-loader": "3.4.2",
-    "eslint": "6.8.0",
-    "eslint-config-prettier": "6.10.0",
-    "eslint-plugin-jest": "23.6.0",
+    "buffer": "^6.0.3",
+    "compression-webpack-plugin": "^11.1.0",
+    "copy-webpack-plugin": "^12.0.2",
+    "crypto-browserify": "^3.12.1",
+    "css-loader": "^7.1.2",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-jest": "^27.9.0",
     "eslint-plugin-mdx": "1.6.8",
-    "eslint-plugin-prettier": "3.1.2",
-    "eslint-plugin-react": "7.18.0",
-    "eslint-plugin-react-hooks": "2.3.0",
-    "file-loader": "5.0.2",
-    "html-webpack-plugin": "3.2.0",
-    "jest": "25.1.0",
-    "prettier": "1.19.1",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.6.3",
+    "jest": "^29.7.0",
+    "os-browserify": "^0.3.0",
+    "path-browserify": "^1.0.1",
+    "prettier": "3.2.5",
+    "process": "^0.11.10",
     "react-docgen-typescript-loader": "3.6.0",
     "semantic-release": "17.0.1",
-    "style-loader": "1.1.3",
-    "typescript": "3.7.5",
-    "webpack": "4.41.5",
+    "stream-browserify": "^3.0.0",
+    "style-loader": "^4.0.0",
+    "typescript": "^5.7.3",
+    "util": "^0.12.5",
+    "vm-browserify": "^1.1.2",
+    "webpack": "^5.99.8",
     "webpack-bundle-analyzer": "^3.6.0",
-    "webpack-cli": "3.3.10",
-    "webpack-dev-server": "3.10.1"
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.2.0"
   },
   "license": "GPL-3.0",
   "babel": {
