+++ src/templates/certificate-of-award/certificate-of-award-template.tsx
@@ -1,8 +1,8 @@
 import React, { FunctionComponent } from "react";
 import "bootstrap/dist/css/bootstrap.css";
-import { TemplateProps } from "@govtechsg/decentralized-renderer-react-components";
+import { TemplateProps } from "@trustvc/decentralized-renderer-react-components";
 import { CertificateOfAward } from "../samples";
-import { css } from "@emotion/core";
+import { css } from "@emotion/react";
 import moeLogo from "./moe-ministry-of-education-singapore.svg";
 import cccLogo from "./citizens-consultative-committees.png";
 import { Page, PageContainer, pageHeight } from "../../components/page";
@@ -11,11 +11,13 @@ const header1Height = "3.5cm";
 const footer1Height = "2cm";
 const content1Height = `calc(${pageHeight} - ${header1Height} - ${footer1Height})`;
 
-export const CertificateOfAwardTemplate: FunctionComponent<TemplateProps<CertificateOfAward>> = ({ document }) => (
-  <PageContainer>
-    <Page>
-      <div
-        css={css`
+export const CertificateOfAwardTemplate: FunctionComponent<TemplateProps<any>> = ({ document }) => {
+  const doc = document as CertificateOfAward;
+  return (
+    <PageContainer>
+      <Page>
+        <div
+          css={css`
           background: linear-gradient(90deg, rgba(255, 255, 255, 1) 0%, rgba(180, 191, 245, 1) 100%);
           display: flex;
           align-items: center;
@@ -27,13 +29,13 @@ export const CertificateOfAwardTemplate: FunctionComponent<TemplateProps<Certifi
             height: 2.5cm;
           }
         `}
-      >
-        <img src={moeLogo} />
-        <img src={cccLogo} />
-      </div>
-      <div
-        className="text-center"
-        css={css`
+        >
+          <img src={moeLogo} />
+          <img src={cccLogo} />
+        </div>
+        <div
+          className="text-center"
+          css={css`
           height: ${content1Height};
           padding: 0 2cm;
           .certificate {
@@ -66,43 +68,44 @@ export const CertificateOfAwardTemplate: FunctionComponent<TemplateProps<Certifi
             margin-bottom: 0.5cm;
           }
         `}
-      >
-        <div className="certificate text-uppercase">{document.name}</div>
-        <div>is presented to</div>
-        <div className="recipient text-uppercase">{document.recipient.name}</div>
-        <div>of</div>
-        <div className="achievement-area text-uppercase">
-          {document.award.achievementArea}
-          <br />
-          in
-          <br />
-          {document.award.instituteName}
-        </div>
-        <div className="information">
-          for demonstrating leadership qualities, service to community and schools,
-          <br />
-          excellence in non-academic activities, and
-          <br />
-          good conduct
-        </div>
-        <div className="signature">
-          <img src={document.signature.signature} />
-          <div className="text-uppercase">
-            {document.signature.name}
+        >
+          <div className="certificate text-uppercase">{doc.name}</div>
+          <div>is presented to</div>
+          <div className="recipient text-uppercase">{doc.recipient.name}</div>
+          <div>of</div>
+          <div className="achievement-area text-uppercase">
+            {doc.award.achievementArea}
             <br />
-            {document.signature.designation}
+            in
+            <br />
+            {doc.award.instituteName}
+          </div>
+          <div className="information">
+            for demonstrating leadership qualities, service to community and schools,
+            <br />
+            excellence in non-academic activities, and
+            <br />
+            good conduct
+          </div>
+          <div className="signature">
+            <img src={doc.signature.signature} />
+            <div className="text-uppercase">
+              {doc.signature.name}
+              <br />
+              {doc.signature.designation}
+            </div>
           </div>
         </div>
-      </div>
-      <div
-        className="footer"
-        css={css`
+        <div
+          className="footer"
+          css={css`
           height: ${footer1Height};
           padding: 0 2cm;
         `}
-      >
-        ABC123456
-      </div>
-    </Page>
-  </PageContainer>
-);
+        >
+          ABC123456
+        </div>
+      </Page>
+    </PageContainer>
+  );
+};
