--- src/templates/govtech-demo-cert/openattestation/OpenAttestationMedia.tsx
@@ -1,15 +1,0 @@
-import React, { FunctionComponent } from "react";
-import { TemplateProps } from "@govtechsg/decentralized-renderer-react-components";
-import "bootstrap/dist/css/bootstrap.css";
-import { GovtechOpencertsTemplateCertificate } from "../../samples";
-
-export const OpenAttestationMedia: FunctionComponent<TemplateProps<GovtechOpencertsTemplateCertificate>> = () => (
-  <div className="embed-responsive embed-responsive-16by9">
-    <iframe
-      id="youtube-vid"
-      className="embed-responsive-item"
-      src="https://www.youtube.com/embed/oskddwGpwUw?autoplay=1"
-      allowFullScreen
-    />
-  </div>
-);
