+++ src/templates/govtech-demo-cert/print.tsx
@@ -1,18 +1,20 @@
 import React, { FunctionComponent } from "react";
-import { TemplateProps } from "@govtechsg/decentralized-renderer-react-components";
+import "bootstrap/dist/css/bootstrap.css";
+import "./common/demo-styles.css";
+import { TemplateProps } from "@trustvc/decentralized-renderer-react-components";
 import { GovtechOpencertsTemplateCertificate } from "../samples";
-import { PrintTemplate as W3cPrintTemplate } from "./w3c/w3cPrint";
-import { GovtechW3cVerifiableCredential } from "../samples/govtech-w3c-template-sample";
-import { OpenAttestationPrint } from "./openattestation/OpenAttestationPrint";
-import { isW3CDocument } from "../../utils/w3c-utils";
+import { CertificateTemplate } from "./certificate";
+import { TranscriptTemplate } from "./transcript";
 
-export const PrintTemplate: FunctionComponent<TemplateProps<GovtechOpencertsTemplateCertificate>> = ({
+export const PrintTemplate: FunctionComponent<TemplateProps<any>> = ({
   document,
   handleObfuscation
-}) => {
-  if (isW3CDocument((document as unknown) as Record<string, unknown>)) {
-    return <W3cPrintTemplate document={(document as unknown) as GovtechW3cVerifiableCredential} />;
-  }
-
-  return <OpenAttestationPrint document={document} handleObfuscation={handleObfuscation} />;
-};
+}) => (
+  <>
+    <CertificateTemplate document={document} handleObfuscation={handleObfuscation} />
+    <div style={{ pageBreakAfter: "always" }} />
+    <TranscriptTemplate document={document} handleObfuscation={handleObfuscation} />
+    <div style={{ pageBreakAfter: "always" }} />
+    <img src="/static/images/yt.png" className="w-100" />
+  </>
+);
