+++ src/templates/imda-demo-cert/common/print-watermark.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent } from "react";
-import { css } from "@emotion/core";
+import { css } from "@emotion/react";
 
 export const PrintWatermark: FunctionComponent = () => (
   <div
