+++ src/templates/imda-demo-cert/common/simple-privacy-filter-banner.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faEdit } from "@fortawesome/free-solid-svg-icons";
-import { css } from "@emotion/core";
+import { css } from "@emotion/react";
 
 interface SimplePrivacyFilterBannerProps {
   /**
