+++ src/templates/imda-demo-cert/media.tsx
@@ -1,9 +1,9 @@
 import React, { FunctionComponent } from "react";
-import { TemplateProps } from "@govtechsg/decentralized-renderer-react-components";
+import { TemplateProps } from "@trustvc/decentralized-renderer-react-components";
 import "bootstrap/dist/css/bootstrap.css";
 import { GovtechOpencertsTemplateCertificate } from "../samples";
 
-export const MediaTemplate: FunctionComponent<TemplateProps<GovtechOpencertsTemplateCertificate>> = () => (
+export const MediaTemplate: FunctionComponent<TemplateProps<any>> = () => (
   <div className="embed-responsive embed-responsive-16by9">
     <iframe
       id="youtube-vid"
