+++ src/utils/w3c-utils.ts
@@ -1,4 +1,6 @@
 export const isW3CDocument = (document: Record<string, unknown>): boolean => {
+  console.log(document);
+  debugger;
   if (!document || !document["@context"]) return false;
 
   const type = document.type;
