+++ webpack.config.js
@@ -17,7 +17,7 @@ module.exports = {
   target: "web",
   output: {
     path: path.resolve(__dirname, "dist"),
-    filename: "[name].[hash:7].js",
+    filename: "[name].[fullhash:7].js",
     publicPath: "/"
   },
   module: {
@@ -44,17 +44,21 @@ module.exports = {
     ]
   },
   plugins: [
-    new webpack.EnvironmentPlugin(["NODE_ENV", "NET"]),
+    new webpack.EnvironmentPlugin({ NODE_ENV: process.env.NODE_ENV, NET: "" }),
+    new webpack.ProvidePlugin({
+      Buffer: ["buffer", "Buffer"],
+      process: "process/browser",
+    }),
     new HtmlWebpackPlugin({
       filename: "index.html",
       template: `${__dirname}/static/index.html`
     }),
     ...(IS_PROD
       ? [
-          new CompressionPlugin({ test: /\.(js|css|html|svg)$/ }),
-          new BrotliPlugin({ test: /\.(js|css|html|svg)$/ }),
-          new CopyWebpackPlugin([{ from: "static/images", to: "static/images" }])
-        ]
+        new CompressionPlugin({ test: /\.(js|css|html|svg)$/ }),
+        new BrotliPlugin({ test: /\.(js|css|html|svg)$/ }),
+        new CopyWebpackPlugin({ patterns: [{ from: "static/images", to: "static/images" }] })
+      ]
       : [])
   ],
   optimization: {
@@ -71,17 +75,15 @@ module.exports = {
 
   // Using cheap-eval-source-map for build times
   // switch to inline-source-map if detailed debugging needed
-  devtool: IS_PROD ? false : "cheap-eval-source-map",
+  devtool: IS_PROD ? false : "eval-cheap-source-map",
 
   devServer: {
     compress: true,
-    disableHostCheck: true,
     historyApiFallback: true,
     hot: true,
-    inline: true,
     port: 3010,
-    stats: {
-      colors: true,
+    allowedHosts: "all",
+    client: {
       progress: true
     }
   },
@@ -90,8 +92,33 @@ module.exports = {
     extensions: [".js", ".ts", ".tsx"],
     modules: ["node_modules", path.resolve(__dirname, "src")],
     alias: {
-      "react-dom": "@hot-loader/react-dom"
-    }
+      react: path.resolve(__dirname, "node_modules/react"),
+      "react-dom": path.resolve(__dirname, "node_modules/react-dom"),
+      "react-dom/client": path.resolve(__dirname, "node_modules/react-dom/client"),
+      "react-dom/server": path.resolve(__dirname, "node_modules/react-dom/server"),
+      "react-dom/test-utils": path.resolve(__dirname, "node_modules/react-dom/test-utils"),
+      "react/jsx-runtime": path.resolve(__dirname, "node_modules/react/jsx-runtime"),
+      "react/jsx-dev-runtime": path.resolve(__dirname, "node_modules/react/jsx-dev-runtime"),
+      scheduler: path.resolve(__dirname, "node_modules/scheduler"),
+    },
+    fallback: {
+      crypto: require.resolve("crypto-browserify"),
+      stream: require.resolve("stream-browserify"),
+      path: require.resolve("path-browserify"),
+      os: require.resolve("os-browserify/browser"),
+      vm: require.resolve("vm-browserify"),
+      buffer: require.resolve("buffer/"),
+      process: require.resolve("process/browser"),
+      assert: require.resolve("assert/"),
+      util: require.resolve("util/"),
+      fs: false,
+    },
   },
+  ignoreWarnings: [
+    {
+      module: /node-bbs-signatures\/lib\/(bbsSignature|bls12381|bls12381toBbs)\.js/,
+      message: /Critical dependency: the request of a dependency is an expression/,
+    },
+  ],
   bail: true
 };
